+++ src/client.rs
@@ -1,6 +1,9 @@
-use std::{ffi::CString, time::Duration};
+use std::{ffi::CString, ptr, time::Duration};
 
-use open62541_sys::{UA_CertificateVerification_AcceptAll, UA_ClientConfig, UA_Client_connect};
+use open62541_sys::{
+    UA_CertificateVerification_AcceptAll, UA_ClientConfig, UA_Client_connect,
+    UA_Client_getEndpoints,
+};
 
 use crate::{ua, DataType as _, Error, Result};
 
@@ -197,6 +200,51 @@ impl ClientBuilder {
         let mut client = self.build();
         client.connect(endpoint_url)?;
         Ok(client)
+    }
+
+    /// Connects to OPC UA server and returns endpoints.
+    ///
+    /// # Errors
+    ///
+    /// This fails when the target server is not reachable.
+    ///
+    /// # Panics
+    ///
+    /// The server URL must not contain any NUL bytes.
+    pub fn get_endpoints(self, server_url: &str) -> Result<ua::Array<ua::EndpointDescription>> {
+        log::info!("Getting endpoints of server {server_url}");
+
+        let server_url = CString::new(server_url).expect("server URL does not contain NUL bytes");
+
+        let mut client = self.build();
+        let endpoint_descriptions: Option<ua::Array<ua::EndpointDescription>>;
+
+        let status_code = ua::StatusCode::new({
+            let mut endpoint_descriptions_size = 0;
+            let mut endpoint_descriptions_ptr = ptr::null_mut();
+            let result = unsafe {
+                UA_Client_getEndpoints(
+                    client.0.as_mut_ptr(),
+                    server_url.as_ptr(),
+                    &mut endpoint_descriptions_size,
+                    &mut endpoint_descriptions_ptr,
+                )
+            };
+            // Wrap array result immediately to not leak memory when leaving function early as with
+            // `?` below.
+            endpoint_descriptions = ua::Array::<ua::EndpointDescription>::from_raw_parts(
+                endpoint_descriptions_size,
+                endpoint_descriptions_ptr,
+            );
+            result
+        });
+        Error::verify_good(&status_code)?;
+
+        let Some(endpoint_descriptions) = endpoint_descriptions else {
+            return Err(Error::internal("expected array of endpoint descriptions"));
+        };
+
+        Ok(endpoint_descriptions)
     }
 
     /// Builds OPC UA client.