+++ CHANGELOG.md
@@ -10,8 +10,8 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Add type `PrivateKey` to wrap private keys when using Mbed TLS.
-- Add ability to fetch remote server certificate with `fetch_server_certificate()`.
 - Add types `ua::SecurityLevel`, `ua::EndpointDescription`, `ua::MessageSecurityMode`.
+- Add ability to get remote server endpoints with `ua::ClientBuilder::get_endpoints()`.
 
 ### Changed